fix: moved auth service
