+++ Weather_App/src/app/app-routing.module.ts
@@ -32,7 +32,7 @@ const routes: Routes = [
     path: 'protected',
     component: ProtectedComponent,
     canActivate: [AuthGuard],
-  },
+  }
 ];
 
 @NgModule({
